# src/sittingface_charts/openapi/services/auth.openapi.yaml
openapi: 3.1.0
info:
  title: auth API
  description: MonkeyBrain auth service API
  version: 1.0.0
paths:
...
  /api/v1/auth/compliance/user-requirements/assess:
    post:
      tags:
      - Auth
      summary: Assess User Requirements
      operationId: assess_user_requirements_api_v1_auth_compliance_user_requirements_assess_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/UserRequirementsAssessmentRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/auth/compliance/traceability-matrix/assess:
    post:
      tags:
      - Auth
      summary: Assess Traceability Matrix
      operationId: assess_traceability_matrix_api_v1_auth_compliance_traceability_matrix_assess_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TraceabilityMatrixRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/auth/compliance/agent/report:
    post:
      tags:
      - Auth
      summary: Save Compliance Agent Report
      operationId: save_compliance_agent_report_api_v1_auth_compliance_agent_report_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceAgentReportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/auth/compliance/user-requirements/assess:
    post:
      tags:
      - Auth-Frontend
      summary: Assess User Requirements
      operationId: assess_user_requirements_api_auth_compliance_user_requirements_assess_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/UserRequirementsAssessmentRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/auth/compliance/traceability-matrix/assess:
    post:
      tags:
      - Auth-Frontend
      summary: Assess Traceability Matrix
      operationId: assess_traceability_matrix_api_auth_compliance_traceability_matrix_assess_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TraceabilityMatrixRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/auth/compliance/agent/report:
    post:
      tags:
      - Auth-Frontend
      summary: Save Compliance Agent Report
      operationId: save_compliance_agent_report_api_auth_compliance_agent_report_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceAgentReportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ComplianceAgentReportRequest:
      properties:
        summary:
          additionalProperties: true
          type: object
          title: Summary
        reports:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Reports
        failed_tests:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Failed Tests
        scope_results:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Scope Results
        message:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Message
        source:
          type: string
          maxLength: 128
          title: Source
          default: n8n
        raw_response:
          additionalProperties: true
          type: object
          title: Raw Response
      type: object
      title: ComplianceAgentReportRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TraceabilityMatrixRequest:
      properties:
        scope:
          type: string
          maxLength: 255
          minLength: 1
          title: Scope
          default: pharma-compliance-proof
        requirement_ids:
          items:
            type: string
          type: array
          title: Requirement Ids
      type: object
      title: TraceabilityMatrixRequest
...
    UserRequirementInput:
      properties:
        requirement_id:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Requirement Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          type: string
          maxLength: 4000
          minLength: 1
          title: Description
        category:
          type: string
          maxLength: 128
          minLength: 1
          title: Category
          default: Compliance
        priority:
          type: string
          pattern: ^(Critical|High|Medium|Low)$
          title: Priority
          default: High
        source:
          type: string
          maxLength: 255
          minLength: 1
          title: Source
          default: User Requirement Specification
        acceptance_criteria:
          items:
            type: string
          type: array
          title: Acceptance Criteria
        linked_controls:
          items:
            type: string
          type: array
          title: Linked Controls
        linked_tests:
          items:
            type: string
          type: array
          title: Linked Tests
        linked_evidence:
          items:
            type: string
          type: array
          title: Linked Evidence
        owner:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Owner
        status:
          type: string
          maxLength: 80
          minLength: 1
          title: Status
          default: Draft
      type: object
      required:
      - title
      - description
      title: UserRequirementInput
    UserRequirementsAssessmentRequest:
      properties:
        scope:
          type: string
          maxLength: 255
          minLength: 1
          title: Scope
          default: pharma-compliance-proof
        requirements:
          items:
            $ref: '#/components/schemas/UserRequirementInput'
          type: array
          title: Requirements
      type: object
      title: UserRequirementsAssessmentRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer